+++ packages/flutter/test/widgets/basic_test.dart
@@ -18,6 +18,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'semantics_tester.dart';
+import 'button_tester.dart';
 import 'utils.dart';
 import 'widgets_app_tester.dart';
